renamed the ui project, this is where all the ui stuff will go
