+++ MiniGames.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33815.320
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MainUI", "MainUI\MainUI.csproj", "{94E22522-6E9E-42C7-85FD-A9501B660D04}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Connect4", "Connect4\Connect4.csproj", "{177D9FD8-4D57-4778-8956-C1F0C4E0AEA6}"
 EndProject
 Global
@@ -13,10 +11,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{94E22522-6E9E-42C7-85FD-A9501B660D04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{94E22522-6E9E-42C7-85FD-A9501B660D04}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{94E22522-6E9E-42C7-85FD-A9501B660D04}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{94E22522-6E9E-42C7-85FD-A9501B660D04}.Release|Any CPU.Build.0 = Release|Any CPU
 		{177D9FD8-4D57-4778-8956-C1F0C4E0AEA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{177D9FD8-4D57-4778-8956-C1F0C4E0AEA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{177D9FD8-4D57-4778-8956-C1F0C4E0AEA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
