+++ UI/Form1.cs
@@ -1,4 +1,4 @@
-namespace MainUI
+namespace UI
 {
     public partial class Form1 : Form
     {
